simulate_animal and simulate_groups needed the travelling time incorporated into the waiting time for exponential draws. Now sim_animal is gillespie again (eg the time at home vs away matches the ttl and ttr fed to function).
as a result, sim-non-ind is getting close to being there, but I think it needs animals to 'go home' more often.

sim-non-ind is not spatially explicit because 'sent home' means to where group vector value is,, not location. I like it this way, but get_times had to change to take this into account when we're evaluating how well the simulators are performing